refactor: js injection / MiniAppWebChromeClient
