+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/display/MiniAppWebChromeClient.kt
@@ -1,0 +1,28 @@
+package com.rakuten.tech.mobile.miniapp.display
+
+import android.content.Context
+import android.webkit.WebChromeClient
+import android.webkit.WebView
+import androidx.annotation.VisibleForTesting
+import java.io.BufferedReader
+
+internal class MiniAppWebChromeClient(context: Context) : WebChromeClient() {
+    @Suppress("TooGenericExceptionCaught", "SwallowedException")
+    @VisibleForTesting
+    internal val bridgeJs = try {
+        val inputStream = context.assets.open("bridge.js")
+        inputStream.bufferedReader().use(BufferedReader::readText)
+    } catch (e: Exception) {
+        null
+    }
+
+    override fun onReceivedTitle(webView: WebView, title: String?) {
+        doInjection(webView)
+        super.onReceivedTitle(webView, title)
+    }
+
+    @VisibleForTesting
+    internal fun doInjection(webView: WebView) {
+        webView.evaluateJavascript(bridgeJs) {}
+    }
+}
