+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/display/RealMiniAppDisplay.kt
@@ -4,21 +4,15 @@ import android.annotation.SuppressLint
 import android.content.Context
 import android.view.View
 import android.view.ViewGroup
-import android.webkit.WebResourceRequest
-import android.webkit.WebResourceResponse
 import android.webkit.WebView
-import android.webkit.WebViewClient
 import android.widget.FrameLayout
 import androidx.annotation.VisibleForTesting
-import androidx.core.net.toUri
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.OnLifecycleEvent
 import androidx.webkit.WebViewAssetLoader
 import com.rakuten.tech.mobile.miniapp.MiniAppDisplay
 import com.rakuten.tech.mobile.miniapp.js.MiniAppMessageBridge
-import java.io.BufferedReader
 import java.io.File
-import java.util.concurrent.atomic.AtomicBoolean
 
 private const val SUB_DOMAIN_PATH = "miniapp"
 private const val MINI_APP_INTERFACE = "MiniAppAndroid"
@@ -47,7 +41,8 @@ internal class RealMiniAppDisplay(
         settings.allowUniversalAccessFromFileURLs = true
         settings.domStorageEnabled = true
         settings.databaseEnabled = true
-        webViewClient = MiniAppWebViewClient(context, getWebViewAssetLoader(), customDomain, customScheme)
+        webViewClient = MiniAppWebViewClient(getWebViewAssetLoader(), customDomain, customScheme)
+        webChromeClient = MiniAppWebChromeClient(context)
 
         loadUrl(getLoadUrl())
     }
@@ -89,48 +84,6 @@ internal class RealMiniAppDisplay(
 
     @VisibleForTesting
     internal fun getLoadUrl() = "$customScheme$SUB_DOMAIN_PATH/index.html"
-}
-
-@VisibleForTesting
-internal class MiniAppWebViewClient(
-    context: Context,
-    @VisibleForTesting internal val loader: WebViewAssetLoader,
-    private val customDomain: String,
-    private val customScheme: String
-) : WebViewClient() {
-    @VisibleForTesting
-    internal var isJsInjected: AtomicBoolean = AtomicBoolean(false)
-
-    @Suppress("TooGenericExceptionCaught", "SwallowedException")
-    @VisibleForTesting
-    internal val bridgeJs = try {
-        val inputStream = context.assets.open("bridge.js")
-        inputStream.bufferedReader().use(BufferedReader::readText)
-    } catch (e: Exception) {
-        null
-    }
-
-    override fun shouldInterceptRequest(
-        view: WebView,
-        request: WebResourceRequest
-    ): WebResourceResponse? {
-        if (request.url != null && request.url.toString().startsWith(customScheme)) {
-            // Do js injection when index is loaded and at least one resource is requested to trigger this.
-            doInjection(view, request.url.toString())
-
-            val interceptUri = request.url.toString().replace(customScheme, customDomain).toUri()
-            return loader.shouldInterceptRequest(interceptUri)
-        }
-        return loader.shouldInterceptRequest(request.url)
-    }
-
-    @VisibleForTesting
-    internal fun doInjection(webView: WebView, urlRequest: String) {
-        if (!isJsInjected.get() && !urlRequest.endsWith("index.html")) {
-            webView.post { webView.evaluateJavascript(bridgeJs) {} }
-            isJsInjected.set(true)
-        }
-    }
 }
 
 internal interface WebViewListener {
